+++ R/vroom_lines.R
@@ -39,10 +39,11 @@ vroom_lines <- function(file, n_max = Inf, skip = 0,
   # be newlines. Ideally this would be "\0", but R doesn't let you have nulls
   # in character vectors.
   out <- vroom_(file, delim = "\1", col_names = "V1", col_types = cols(col_character()),
-    id = NULL, skip = skip, col_select = col_select, na = character(), quote = "",
-    trim_ws = FALSE, escape_double = FALSE, escape_backslash = FALSE, comment = "",
-    locale = locale, guess_max = 0, n_max = n_max, altrep = vroom_altrep(altrep),
-    num_threads = num_threads, progress = progress
+    id = NULL, skip = skip, col_select = col_select, name_repair = "minimal",
+    na = character(), quote = "", trim_ws = FALSE, escape_double = FALSE,
+    escape_backslash = FALSE, comment = "", locale = locale, guess_max = 0,
+    n_max = n_max, altrep = vroom_altrep(altrep), num_threads = num_threads,
+    progress = progress
   )
 
   out[[1]]