+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * `vroom()` now handles cases when `id` is set and a column in skipped (#237)
 
+* `vroom()` now supports column selections when there are some empty column names (#238)
+
 # vroom 1.2.1
 
 * `vroom()` now writes the column names if given an input with no rows (#213)